add support for Angular 13
