+++ .gitignore
@@ -1,21 +1,19 @@
 # See http://help.github.com/ignore-files/ for more about ignoring files.
 
-# compiled output
+# Compiled output
 /tmp
 /dist
 /out-tsc
 # Only exists if Bazel was run
 /bazel-out
 
-# dependencies
+# Node
 /node_modules
-
-# profiling files
-chrome-profiler-events.json
-speed-measure-plugin.json
+npm-debug.log
+yarn-error.log
 
 # IDEs and editors
-/.idea
+.idea/
 .project
 .classpath
 .c9/
@@ -31,16 +29,15 @@ speed-measure-plugin.json
 !.vscode/extensions.json
 .history/*
 
-# misc
+# Miscellaneous
+/.angular
 /.sass-cache
 /connect.lock
 /coverage
 /libpeerconnection.log
-npm-debug.log
-yarn-error.log
 testem.log
 /typings
 
-# System Files
+# System files
 .DS_Store
 Thumbs.db
