+++ angular.json
@@ -1,6 +1,9 @@
 {
   "$schema": "./node_modules/@angular/cli/lib/config/schema.json",
   "version": 1,
+  "cli": {
+    "packageManager": "yarn"
+  },
   "newProjectRoot": "projects",
   "projects": {
     "ngx-translate-mock": {
@@ -10,28 +13,27 @@
       "prefix": "lib",
       "architect": {
         "build": {
-          "builder": "@angular-devkit/build-ng-packagr:build",
+          "builder": "@angular-devkit/build-angular:ng-packagr",
           "options": {
-            "tsConfig": "projects/ngx-translate-mock/tsconfig.lib.json",
             "project": "projects/ngx-translate-mock/ng-package.json"
-          }
-        },
-        "test": {
-          "builder": "@angular-devkit/build-angular:karma",
-          "options": {
-            "main": "projects/ngx-translate-mock/src/test.ts",
-            "tsConfig": "projects/ngx-translate-mock/tsconfig.spec.json",
-            "karmaConfig": "projects/ngx-translate-mock/karma.conf.js"
-          }
+          },
+          "configurations": {
+            "production": {
+              "tsConfig": "projects/ngx-translate-mock/tsconfig.lib.prod.json"
+            },
+            "development": {
+              "tsConfig": "projects/ngx-translate-mock/tsconfig.lib.json"
+            }
+          },
+          "defaultConfiguration": "production"
         },
         "lint": {
-          "builder": "@angular-devkit/build-angular:tslint",
+          "builder": "@angular-eslint/builder:lint",
           "options": {
-            "tsConfig": [
-              "projects/ngx-translate-mock/tsconfig.lib.json",
-              "projects/ngx-translate-mock/tsconfig.spec.json"
-            ],
-            "exclude": ["**/node_modules/**"]
+            "lintFilePatterns": [
+              "projects/ngx-translate-mock/**/*.ts",
+              "projects/ngx-translate-mock/**/*.html"
+            ]
           }
         }
       }
