+++ package.json
@@ -6,6 +6,7 @@
   "scripts": {
     "ng": "ng",
     "build": "ng build && yarn copy:readme",
+    "watch": "ng build --watch --configuration development",
     "lint": "ng lint",
     "pretty": "prettier --list-different --config ./.prettierrc \"**/*.@(js|ts|scss|md|json|yml|html)\"",
     "pretty:write": "prettier --write --config ./.prettierrc \"**/*.@(js|ts|scss|md|json|yml|html|template)\"",
@@ -14,27 +15,25 @@
   "private": true,
   "dependencies": {},
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.1102.6",
-    "@angular-devkit/build-ng-packagr": "0.1002.0",
-    "@angular/cli": "11.2.6",
-    "@angular/common": "11.2.7",
-    "@angular/compiler": "11.2.7",
-    "@angular/compiler-cli": "11.2.7",
-    "@angular/core": "11.2.7",
-    "@angular/language-service": "11.2.7",
-    "@ngx-translate/core": "13.0.0",
-    "@types/jasmine": "3.6.9",
-    "@types/jasminewd2": "2.0.8",
-    "@types/node": "14.14.37",
-    "codelyzer": "6.0.1",
+    "@angular-devkit/build-angular": "13.2.2",
+    "@angular-eslint/builder": "13.0.1",
+    "@angular-eslint/eslint-plugin": "13.0.1",
+    "@angular-eslint/schematics": "13.0.1",
+    "@angular/cli": "13.2.2",
+    "@angular/compiler": "13.2.2",
+    "@angular/compiler-cli": "13.2.2",
+    "@angular/core": "13.2.2",
+    "@ngx-translate/core": "14.0.0",
+    "@types/node": "12.11.1",
+    "@typescript-eslint/eslint-plugin": "5.3.0",
+    "@typescript-eslint/parser": "5.3.0",
     "copyfiles": "2.4.1",
-    "ng-packagr": "11.2.4",
-    "prettier": "2.2.1",
-    "rxjs": "6.6.7",
-    "ts-node": "9.1.1",
-    "tsickle": "0.39.1",
-    "tslib": "2.1.0",
-    "tslint": "6.1.3",
-    "typescript": "4.1.5"
+    "eslint": "8.2.0",
+    "ng-packagr": "13.0.0",
+    "prettier": "2.5.1",
+    "rxjs": "6.6.3",
+    "tslib": "2.3.0",
+    "typescript": "4.4.4",
+    "zone.js": "0.11.4"
   }
 }
