+++ projects/ngx-translate-mock/package.json
@@ -4,9 +4,8 @@
   "license": "MIT",
   "version": "0.2.0",
   "peerDependencies": {
-    "@angular/common": ">=6",
-    "@angular/core": ">=6",
-    "@ngx-translate/core": ">=10",
+    "@angular/core": ">=13",
+    "@ngx-translate/core": "^14.0.0",
     "rxjs": ">=6"
   }
 }
