+++ projects/ngx-translate-mock/src/lib/ngx-translate-mock.module.ts
@@ -1,103 +1,14 @@
 /**
- * Copyright (c) 2020 Hetzner Cloud GmbH
+ * Copyright (c) 2022 Hetzner Cloud GmbH
  *
  * SPDX-License-Identifier: MIT
  */
 
-import {
-  AfterViewChecked,
-  ClassProvider,
-  Directive,
-  ElementRef,
-  Injectable,
-  Input,
-  NgModule,
-  Pipe,
-  PipeTransform,
-} from '@angular/core';
-import { LangChangeEvent, TranslateService } from '@ngx-translate/core';
-import { BehaviorSubject, Observable, of, Subject } from 'rxjs';
-
-export const TRANSLATED_STRING: string = 'i18n';
-
-@Injectable()
-export class TranslateServiceMock {
-    onLangChangeSubject: Subject<LangChangeEvent> = new Subject();
-    onTranslationChangeSubject: Subject<string> = new Subject();
-    onDefaultLangChangeSubject: Subject<string> = new Subject();
-    isLoadedSubject: BehaviorSubject<boolean> = new BehaviorSubject(true);
-
-    onLangChange: Observable<
-        LangChangeEvent
-    > = this.onLangChangeSubject.asObservable();
-    onTranslationChange: Observable<
-        string
-    > = this.onTranslationChangeSubject.asObservable();
-    onDefaultLangChange: Observable<
-        string
-    > = this.onDefaultLangChangeSubject.asObservable();
-    isLoaded: Observable<boolean> = this.isLoadedSubject.asObservable();
-
-    currentLang: string;
-
-    languages: string[] = ['de'];
-
-    get(content: string): Observable<string> {
-        return of(TRANSLATED_STRING + content);
-    }
-
-    use(lang: string): void {
-        this.currentLang = lang;
-        this.onLangChangeSubject.next({ lang } as LangChangeEvent);
-    }
-
-    addLangs(langs: string[]): void {
-        this.languages = [...this.languages, ...langs];
-    }
-
-    getBrowserLang(): string {
-        return '';
-    }
-
-    getLangs(): string[] {
-        return this.languages;
-    }
-
-    // tslint:disable-next-line:no-any
-    getTranslation(): Observable<any> {
-        return of({});
-    }
-
-    instant(key: string | string[], interpolateParams?: object): string {
-        return TRANSLATED_STRING + key.toString();
-    }
-
-    setDefaultLang(lang: string): void {
-        this.onDefaultLangChangeSubject.next(lang);
-    }
-}
-
-@Pipe({ name: 'translate' })
-export class TranslateMockPipe implements PipeTransform {
-    transform(text: string): string {
-        return !text ? TRANSLATED_STRING : `${text}-${TRANSLATED_STRING}`;
-    }
-}
-
-@Directive({
-    // tslint:disable-next-line:directive-selector
-    selector: '[translate]',
-})
-// tslint:disable:no-any
-export class TranslateMockDirective implements AfterViewChecked {
-    @Input()
-    translateParams: any;
-    constructor(private readonly _element: ElementRef) {}
-
-    ngAfterViewChecked(): void {
-        this._element.nativeElement.innerText += TRANSLATED_STRING;
-    }
-}
+import { ClassProvider, NgModule } from '@angular/core';
+import { TranslateService } from '@ngx-translate/core';
+import { TranslateServiceMock } from './ngx-translate-mock.service';
+import { TranslateMockDirective } from './ngx-translate-mock.directive';
+import { TranslateMockPipe } from './ngx-translate-mock.pipe';
 
 export const TRANSLATE_SERVICE_MOCK: ClassProvider = {
     provide: TranslateService,
