+++ projects/ngx-translate-mock/src/lib/util.ts
@@ -1,0 +1,1 @@
+export const TRANSLATED_STRING: string = 'i18n';
