+++ projects/ngx-translate-mock/src/public-api.ts
@@ -2,4 +2,8 @@
  * Public API Surface of ngx-translate-mock
  */
 
+export * from './lib/ngx-translate-mock.directive';
 export * from './lib/ngx-translate-mock.module';
+export * from './lib/ngx-translate-mock.pipe';
+export * from './lib/ngx-translate-mock.service';
+export * from './lib/util';
