--- projects/ngx-translate-mock/tslint.json
@@ -1,7 +1,0 @@
-{
-  "extends": "../../tslint.json",
-  "rules": {
-    "directive-selector": [true, "attribute", "lib", "camelCase"],
-    "component-selector": [true, "element", "lib", "kebab-case"]
-  }
-}
